requirements.txt
Flask
opencv-python-headless
Pillow
numpy
Werkzeug
gunicorn